+++ src/components/time/dateSelect/ThisWeek.tsx
@@ -1,45 +1,6 @@
 import styled from 'styled-components';
 import Typo from '../../../styles/typo/typo';
-
-// 요일 배열
-const daysOfWeek = ['일', '월', '화', '수', '목', '금', '토'];
-
-interface DateInfo {
-	day: number;
-	dayOfWeek: string;
-}
-
-interface DayOfWeekProps {
-	$dayOfWeek: string;
-}
-
-// 토요일은 파란색, 일요일은 빨간색
-const getColorByDay = ($dayOfWeek: string): string => {
-	switch ($dayOfWeek) {
-		case '토':
-			return '#66B3FC';
-		case '일':
-			return '#F84C4C';
-		default:
-			return '#373737';
-	}
-};
-
-/** 이번주 날짜 포맷 함수 */
-const formatDate = (date: Date): DateInfo => {
-	const day = date.getDate();
-	const dayOfWeek = daysOfWeek[date.getDay()];
-	return { day, dayOfWeek };
-};
-
-/** 오늘 날짜 포맷 함수 */
-const formatToday = (date: Date): string => {
-	const year = date.getFullYear();
-	const month = date.getMonth() + 1;
-	const day = date.getDate();
-	const dayOfWeek = daysOfWeek[date.getDay()];
-	return `${year}.${month}.${day} (${dayOfWeek})`;
-};
+import { getColorByDay, formatDate, formatToday } from '../../../utils/time/dateUtils';
 
 function ThisWeek() {
 	const today = new Date();
@@ -48,6 +9,7 @@ function ThisWeek() {
 	const weekDates = Array.from({ length: 7 }, (_, i) => {
 		const date = new Date(today);
 		date.setDate(today.getDate() + i + 1);
+		// util에서 요일 불러오기
 		const { day, dayOfWeek } = formatDate(date);
 		return {
 			day,
@@ -61,6 +23,7 @@ function ThisWeek() {
 	return (
 		<>
 			<TodayWrapper>
+				{/* util에서 오늘 날짜 불러오기 */}
 				<Typo.Title.Title4M16>{formatToday(today)} </Typo.Title.Title4M16>
 				<Blue>오늘</Blue>
 			</TodayWrapper>
@@ -68,14 +31,11 @@ function ThisWeek() {
 			<WeekDateContainer>
 				{weekDates.map((weekDate) => (
 					<WeekDateWrapper key={weekDate.uniqueKey}>
-						<WeekDateDay>{weekDate.day}</WeekDateDay>
+						<Typo.Head.Head1SB17>{weekDate.day}</Typo.Head.Head1SB17>
 						{weekDate.label === '내일' ? (
-							<LabelRed>{weekDate.label}</LabelRed>
+							<Red>{weekDate.label}</Red>
 						) : (
-							<>
-								<DayOfWeek $dayOfWeek={weekDate.dayOfWeek}>{weekDate.dayOfWeek}</DayOfWeek>
-								<Typo.Body.Body5M13>{weekDate.label}</Typo.Body.Body5M13>
-							</>
+							<DayOfWeek $dayOfWeek={weekDate.dayOfWeek}>{weekDate.dayOfWeek}</DayOfWeek>
 						)}
 					</WeekDateWrapper>
 				))}
@@ -84,24 +44,24 @@ function ThisWeek() {
 	);
 }
 
-const WeekDateDay = styled(Typo.Head.Head1SB17)`
-	margin-bottom: 1rem;
-`;
-
-const LabelRed = styled(Typo.Body.Body5M13)`
-	color: ${(props) => props.theme.Color.Point};
-`;
-
 const TodayWrapper = styled.div`
 	display: flex;
 	align-items: center;
 	justify-content: center;
-	padding: 3rem 0 0.4rem;
+	margin-bottom: 1rem;
+	padding: 2rem;
 
 	color: ${(props) => props.theme.GreyScale.BG};
 `;
 
+interface DayOfWeekProps {
+	$dayOfWeek: string;
+}
+
+/* util에서 요일 색상 불러오기 */
 const DayOfWeek = styled(Typo.Body.Body5M13)<DayOfWeekProps>`
+	margin-top: 1rem;
+
 	color: ${({ $dayOfWeek }) => getColorByDay($dayOfWeek)};
 `;
 
@@ -121,6 +81,12 @@ const Blue = styled(Typo.Title.Title4M16)`
 	padding-left: 0.4rem;
 
 	color: ${(props) => props.theme.Color.SkyBlue};
+`;
+
+const Red = styled(Typo.Body.Body5M13)`
+	margin-top: 1rem;
+
+	color: ${(props) => props.theme.Color.Point};
 `;
 
 export default ThisWeek;
